# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gaming House Station API — PROPOSAL",
    "version": "0.1.0",
    "description": "Phase 1 proposal for partner review. No production endpoints are asserted to exist. Station scope only; no cashier powers. Canonical API version 1.",
    "contact": {
      "name": "Gaming House / Padel House integration owners"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:4317/gaming/v1",
      "description": "Development simulator address reserved for Phase 4; not running in Phase 1. Real adapter requires configured HTTPS."
    }
  ],
  "security": [
    {
      "stationBearer": []
    }
  ],
...
  "paths": {
    "/devices/enroll": {
      "post": {
        "operationId": "enrollDevice",
        "summary": "Redeem a one-time partner-issued station pairing code",
        "tags": [
          "Enrollment"
        ],
        "description": "PROPOSED, not an existing Padel House route. See docs/partner-handoff.md for transactional, timing, and retry semantics.",
        "security": [],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Stable UUID for one logical mutation; same key + different canonical payload returns 409. Reuse identical body on retry.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Accepted response. A launch permit is not proof that a game is running or billing has started.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid message.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Credential, renewal token, or pairing code invalid/expired/revoked.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Station scope or authorization rejected.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Session, revision, sequence, or idempotency conflict; reconcile before a new action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "410": {
            "description": "Expired authorization, attempt, or command.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Back off using Retry-After.",
            "headers": {
              "Retry-After": {
                "description": "Seconds before retry.",
                "schema": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 9007199254740991
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Transient failure; retry with stable identifiers and backoff.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "stationBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Per-station revocable opaque bearer token over HTTPS; never a staff cookie. Enrollment uses a one-time code; renewal uses its dedicated scoped secret."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "api_version": {
            "type": "string",
            "const": "1"
          },
          "request_id": {
            "type": "string",
            "format": "uuid"
          },
          "station_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ]
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ]
          },
          "server_time": {
            "type": "string",
            "format": "date-time",
            "pattern": "Z$",
            "description": "UTC RFC3339 timestamp ending in Z. Server validates device observations; device wall time is not financial authority."
          },
          "code": {
            "type": "string",
            "enum": [
              "INVALID_REQUEST",
              "UNAUTHENTICATED",
              "CREDENTIAL_REVOKED",
              "FORBIDDEN_STATION",
              "PAIRING_EXPIRED",
              "AUTHORIZATION_REQUIRED",
              "AUTHORIZATION_EXPIRED",
              "STATION_UNAVAILABLE",
              "SESSION_MISMATCH",
              "REVISION_CONFLICT",
              "IDEMPOTENCY_CONFLICT",
              "ATTEMPT_EXPIRED",
              "ATTEMPT_CANCELLED",
              "GAME_NOT_ALLOWED",
              "STALE_EVENT",
              "SEQUENCE_CONFLICT",
              "COMMAND_EXPIRED",
              "UNSUPPORTED_COMMAND",
              "RATE_LIMITED",
              "TEMPORARILY_UNAVAILABLE"
            ]
          },
          "message": {
            "type": "string",
            "description": "Safe diagnostic without secrets or customer credentials.",
            "maxLength": 512
          },
          "retryable": {
            "type": "boolean"
          },
          "retry_after_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1,
                "maximum": 9007199254740991
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "api_version",
          "request_id",
          "station_id",
          "session_id",
          "server_time",
          "code",
          "message",
          "retryable",
          "retry_after_seconds"
        ]
      },
      "EnrollRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "api_version": {
            "type": "string",
            "const": "1"
          },
          "request_id": {
            "type": "string",
            "format": "uuid"
          },
          "sent_at": {
            "type": "string",
            "format": "date-time",
            "pattern": "Z$",
            "description": "UTC RFC3339 timestamp ending in Z. Server validates device observations; device wall time is not financial authority."
          },
          "pairing_code": {
            "type": "string",
            "description": "One-use code entered through protected setup.",
            "minLength": 8,
            "maxLength": 128
          },
          "installation_id": {
            "type": "string",
            "format": "uuid"
          },
          "agent_version": {
            "type": "string",
            "description": "Installed application version."
          },
          "station_label": {
            "type": "string",
            "description": "Requested label, backend assigns identity.",
            "pattern": "^PC-[0-9]{2,4}$"
          }
        },
        "required": [
          "api_version",
          "request_id",
          "sent_at",
          "pairing_code",
          "installation_id",
          "agent_version",
          "station_label"
        ]
      },
      "Credentials": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "credential_id": {
            "type": "string",
            "format": "uuid"
          },
          "access_token": {
            "type": "string",
            "description": "Opaque secret. Never log or expose to the renderer.",
            "minLength": 32,
            "maxLength": 4096
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "pattern": "Z$",
            "description": "UTC RFC3339 timestamp ending in Z. Server validates device observations; device wall time is not financial authority."
          },
          "renewal_token": {
            "type": "string",
            "description": "Opaque secret. Never log or expose to the renderer.",
            "minLength": 32,
            "maxLength": 4096
          },
          "renewal_expires_at": {
            "type": "string",
            "format": "date-time",
            "pattern": "Z$",
            "description": "UTC RFC3339 timestamp ending in Z. Server validates device observations; device wall time is not financial authority."
          }
        },
        "required": [
          "credential_id",
          "access_token",
          "expires_at",
          "renewal_token",
          "renewal_expires_at"
        ]
      },
      "EnrollResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "api_version": {
            "type": "string",
            "const": "1"
          },
          "request_id": {
            "type": "string",
            "format": "uuid"
          },
          "station_id": {
            "type": "string",
            "format": "uuid"
          },
          "server_time": {
            "type": "string",
            "format": "date-time",
            "pattern": "Z$",
            "description": "UTC RFC3339 timestamp ending in Z. Server validates device observations; device wall time is not financial authority."
          },
          "station_label": {
            "type": "string",
            "description": "Backend-approved station label."
          },
          "credentials": {
            "$ref": "#/components/schemas/Credentials"
          }
        },
        "required": [
          "api_version",
          "request_id",
          "station_id",
          "server_time",
          "station_label",
          "credentials"
        ]
      },
...
    }
  }
}